email validator updated
