+++ src/main/java/com/student/studentRegisterLogin/registration/EmailValidator.java
@@ -9,6 +9,9 @@ public class EmailValidator implements Predicate<String> {
     @Override
     public boolean test(String s) {
 //        TODO: Regex to validate email
-        return true;
+if(s.contains("@")){
+    return true;
+}
+        return false;
     }
 }
